Updated Peripheral Code
This code is designed to be integrated into the mainline firmware. It enables the STM32 to communicate over I2C2 using pins PB10 (SCL) and PB11 (SDA), acting as an I2C slave. It can receive command bytes from the controller to return sensor data (as floats) and also supports receiving float values from the controller for runtime configuration.